requirements.txt
argon2-cffi==20.1.0
Babel==2.8.0
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.3
django-filter==2.3.0
django-mathfilters==1.0.0
django-phonenumber-field==4.0.0
django-rest-swagger==2.2.0
django-taggit==1.3.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
Faker==4.0.1
gunicorn==20.0.4
httpie==2.0.0
idna==2.9
importlib-metadata==1.7.0
itypes==1.1.0
Jinja2==2.11.1
Markdown==3.2.2
MarkupSafe==1.1.1
mysqlclient==1.4.6
openapi-codec==1.3.2
Pillow==7.2.0
psycopg2==2.8.4
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.25.8
zipp==3.1.0
